nvim 0.10 updates
This brings some of the keybindings in line with new defaults. It also
updates show_line_diagnostics to open_float...I do not thing the former
exists any more? Lastly, fixes [d and ]d mappings as the functions moved
a lot earlier in 0.6.0 and I never noticed. I really like C-J/C-K, but that
is for ale, which might still be useful but idk any more